cleaned up comments and renamed deque_copy
